reorganize application structure
