+++ public/410b88da-0ed1-4749-903f-5e76c24e2e5f/ramp-config/en/OilSandsDeposits.json
