+++ public/ea24000c-7dc3-49a9-baac-c55d28dcaeb9/README.md
@@ -1,0 +1,1 @@
+Placeholder for ea24000c-7dc3-49a9-baac-c55d28dcaeb9 (Ethylene Glycol) Data
